requirements.txt
python-telegram-bot[ext]
flask
razorpay
uuid
setuptools